fixed a bug and introduced a new [plugin] section in the confile
The Bug was, that all messages sent to mumble from irc started with
"mumble: " instead of "irc: " this is now fixed.
Also this text is now coloured and so much nicer to read.
The new [plugin] section make it much easier to include plugins
and provides better documentation and overview.