Make gtfo command drop connection immediately
Always send QUIT command when dropping IRC connection.
Otherwise, it'll open a new connection to IRC while the
the IRC network still thinks the old connection is active
(even if the TCP connection is already closed). This
prevents the new one from acquiring the proper nick or
joining the channel.